renamed context.js file to themeContext.js to easily identify the file
